+++ src/models/index.ts
@@ -28,11 +28,9 @@ export { SortType } from './searchservice/request/SortType';
 export { Direction } from './searchservice/request/Direction';
 
 export { AppliedQueryFilter } from './searchservice/response/AppliedQueryFilter';
-export {
-  DirectAnswer,
-  FieldValueDirectAnswer,
-  FeaturedSnippetDirectAnswer
-} from './searchservice/response/DirectAnswer';
+export { DirectAnswer } from './searchservice/response/DirectAnswer';
+export { FieldValueDirectAnswer } from './searchservice/response/FieldValueDirectAnswer';
+export { FeaturedSnippetDirectAnswer } from './searchservice/response/FeaturedSnippetDirectAnswer';
 export { DirectAnswerType } from './searchservice/response/DirectAnswerType';
 export { Facet, FacetOption } from './searchservice/request/Facet';
 export { DisplayableFacet, DisplayableFacetOption } from './searchservice/response/DisplayableFacet';
