+++ src/transformers/searchservice/createDirectAnswer.ts
@@ -1,8 +1,9 @@
-import { FeaturedSnippetDirectAnswer, FieldValueDirectAnswer } from '../../models/searchservice/response/DirectAnswer';
+import { FeaturedSnippetDirectAnswer } from '../../models/searchservice/response/FeaturedSnippetDirectAnswer';
+import { FieldValueDirectAnswer } from '../../models/searchservice/response/FieldValueDirectAnswer';
 import { DirectAnswerType } from '../../models/searchservice/response/DirectAnswerType';
 import { ResultsFactory } from './ResultsFactory';
 
-export function createDirectAnswer(data: any): FeaturedSnippetDirectAnswer | FieldValueDirectAnswer | undefined {
+export function createDirectAnswer(data: any): FeaturedSnippetDirectAnswer | FieldValueDirectAnswer {
   const isFieldValueDirectAnswer = data?.type === DirectAnswerType.FieldValue;
   const isFeaturedSnippetDirectAnswer = data?.type === DirectAnswerType.FeaturedSnippet;
   if (isFieldValueDirectAnswer) {
@@ -24,5 +25,7 @@ export function createDirectAnswer(data: any): FeaturedSnippetDirectAnswer | Fie
       verticalKey: data.relatedItem.verticalConfigId,
       snippet: data.answer.snippet,
     };
+  } else {
+    throw new Error('The Answers API returned an unknown direct answer type');
   }
 }
